Refactor all dependencies into systems
