+++ app/views/play/cinematic/Command/CinematicParser.js
@@ -1,8 +1,48 @@
-import { getLeftCharacterThangTypeSlug, getRightCharacterThangTypeSlug } from '../../../../schemas/selectors/cinematic'
-
 export const parseShot = (shot, systems) => {
   return parseSetup(shot, systems)
 }
+
+/**
+ * @typedef {import('../../../../schemas/selectors/cinematic').Cinematic} Cinematic
+ */
+
+/**
+ * @typedef {import('../../../../schemas/selectors/cinematic').Shot} Shot
+ */
+
+/**
+ * @typedef {import('../../../../schemas/selectors/cinematic').ShotSetup} ShotSetup
+ */
+
+/**
+ * @typedef {import('../../../../schemas/selectors/cinematic').DialogNode} DialogNode
+ */
+
+/**
+ * @typedef {Object} CommandTuple
+ * @property {AbstractCommand[]} commands - The commands to be run to execute this dialogNode or shotSetup.
+ * @property {AbstractCommand[]} cleanupCommands - commands that can be used to cleanup prior commands.
+ */
+
+/**
+ * Interface for classes that represent a System.
+ *
+ * @interface System
+ */
+
+/**
+ * @function
+ * @name System#parseSetupShot
+ * @param {Shot} - The data of the current shot.
+ * @returns {CommandTuple} - commands are run immediately. cleanupCommands will be run at the end of the shot.
+ */
+
+/**
+ * @function
+ * @name System#parseDialogNode
+ * @param {DialogNode} - The dialogNode data.
+ * @returns {CommandTuple} - commands are run immediately. cleanupCommands get run just before next dialogNode runs.
+ */
 
 /**
  * @typedef {Object} Systems
@@ -15,31 +55,10 @@ export const parseShot = (shot, systems) => {
  * @param {Systems} systems
  */
 const parseSetup = (shot, systems) => {
-  const { cinematicLankBoss, loader } = systems
-  const setupCommands = []
+  const { cinematicLankBoss } = systems
+  let setupCommands = []
 
-  const leftCharSlug = getLeftCharacterThangTypeSlug(shot)
-
-  if (leftCharSlug) {
-    const { slug, enterOnStart, position } = leftCharSlug
-    cinematicLankBoss.addLank('left', loader.getThangType(slug), systems)
-    if (enterOnStart) {
-      setupCommands.push(cinematicLankBoss.moveLankCommand('left', position))
-    } else {
-      cinematicLankBoss.moveLank('left', position)
-    }
-  }
-
-  const rightCharSlug = getRightCharacterThangTypeSlug(shot)
-  if (rightCharSlug) {
-    const { slug, enterOnStart, position } = rightCharSlug
-    cinematicLankBoss.addLank('right', loader.getThangType(slug), systems)
-    if (enterOnStart) {
-      setupCommands.push(cinematicLankBoss.moveLankCommand('right', position))
-    } else {
-      cinematicLankBoss.moveLank('right', position)
-    }
-  }
+  setupCommands = setupCommands.concat(cinematicLankBoss.parseSetupShot(shot).commands || [])
 
   return setupCommands
 }
